financial_kg_system/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
neo4j==5.14.0
openpyxl==3.1.2
networkx==3.2.1
redis==5.0.1
pandas==2.1.3
numpy==1.26.2
python-multipart==0.0.6
pyyaml==6.0.1